Fixing some tests. Suppressing others
